[fix] наконец-то добавлено все необходимые таблички в проект.

+++ AccountingScholarships.Domain/Entities/university/EduEmployeePositions.cs
@@ -13,7 +13,7 @@ namespace AccountingScholarships.Domain.Entities.university
         public DateOnly? EndedOn { get; set; }
         public string LastUpdatedBy { get; set; }
         public DateTime LastUpdatedOn { get; set; }
-        public float? Rate { get; set; }
+        public double? Rate { get; set; }
         public bool? IsMainPosition { get; set; }
         public int? HrOrderId { get; set; }
